+++ TeamCode/src/main/java/Teleoperation/IntoTheDeep_TeleOp.java
@@ -1,6 +1,9 @@
-   package org.firstinspires.ftc.teamcode;
+   package Teleoperation;
    import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
    import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
+
+   import org.firstinspires.ftc.teamcode.Assemblies.Robot;
+
 
 // This file is the main TeleOp file.
