+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AdvancedAutoObservationRung.java
@@ -1,10 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Autonomous;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.IMU;
 import com.qualcomm.robotcore.util.ElapsedTime;
 import com.qualcomm.robotcore.util.Range;
@@ -12,9 +12,9 @@ import com.qualcomm.robotcore.util.Range;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
 
-@Autonomous(name="AutoBasketSpike", group="Robot")
-//@Disabled
-public class AutoBasketSpike extends LinearOpMode {
+@Autonomous(name="AdvancedAutoObservationRung", group="Robot")
+@Disabled
+ public class AdvancedAutoObservationRung extends LinearOpMode {
 
         /* Declare OpMode members. */
         private DcMotor leftFrontDrive   = null;
@@ -24,7 +24,7 @@ public class AutoBasketSpike extends LinearOpMode {
         private IMU imu         = null;
         private final ElapsedTime Time = new ElapsedTime();
         // Control/Expansion Hub IMU
-        LimbArm arm;
+
         private double          headingError  = 0;
 
         // These variable are declared here (as class members) so they can be updated in various methods,
@@ -111,7 +111,6 @@ public class AutoBasketSpike extends LinearOpMode {
         @Override
         public void runOpMode() {
 
-            arm = new LimbArm(hardwareMap, telemetry);
             // Initialize the drive system variables.
             leftFrontDrive  = hardwareMap.get(DcMotor.class, "leftFront");
             rightFrontDrive = hardwareMap.get(DcMotor.class, "rightFront");
@@ -158,34 +157,41 @@ public class AutoBasketSpike extends LinearOpMode {
             } // end of while loop
 
             // Set the encoders for closed loop speed control, and reset the heading.
-            // BEGIN AUTO CODE
+            // BEGIN AUTO CODE //
 
             driveStraight(TURN_SPEED, 22, 0);
             Wait(1);
             // this is where the code for the grabbing mechanism will go
-            driveStraight(TURN_SPEED, -15, 0);
-            // Wait(1);
-            StrafeRobot(TURN_SPEED, 6, 0);
-            // Wait(1);
-            turnToHeading(TURN_SPEED, -88.0);
-            // Wait(1);
-            driveStraight(TURN_SPEED, 62, -88.0);
-            // Wait(1);
-            turnToHeading(TURN_SPEED, 180.0);
-            // Wait(1);
-            driveStraight(TURN_SPEED, 13, 180.0);
+            turnToHeading(TURN_SPEED, -90.0);
+           // Wait(1);
+            driveStraight(TURN_SPEED, 24, -90.0);
+           // Wait(1);
+            StrafeRobot(TURN_SPEED, -16, -90.0);
+           // Wait(1);
+            driveStraight(TURN_SPEED, 7, 180);
+           // Wait(1);
+            turnToHeading(TURN_SPEED, 180);
+           // Wait(1);
+            driveStraight(TURN_SPEED, 26, 180);
             Wait(1);
-            arm.ExtendAutoArm(1900);
-            Wait(0.2);
-            arm.armRotateAuto(-500);
-            Wait(0.2);
-            arm.armRotateAuto(-1500);
-//            Wait(0.2);
-//            arm.armRotateAuto(-1500);
-//            Wait(0.2);
-            //arm.armRotateAuto(-2000);
-            Wait(4);
-            // This is where code for raising the arm to get a lvl. 1 ascent
+            // code for the grabbing mechanism... again
+            driveStraight(TURN_SPEED, -6, 180);
+           // Wait(1);
+            StrafeRobot(TURN_SPEED, 20, 180);
+           // Wait(1);
+            turnToHeading(TURN_SPEED, 90);
+           // Wait(1);
+            driveStraight(TURN_SPEED, 16, 90);
+           // Wait(1);
+            turnToHeading(TURN_SPEED, 0);
+           // Wait(1);
+            driveStraight(TURN_SPEED, 10, 0);
+            Wait(1);
+            //This is for the grabbing mechanism
+            driveStraight(TURN_SPEED, -22, 0);
+           // Wait(1);
+            StrafeRobot(TURN_SPEED, 29, 0);
+            Wait(1);
 
 
             // 0 for heading is whatever the robot's original position is. 90 goes to the left of 0, and negative 90 goes to the right of the 0.
@@ -543,7 +549,11 @@ public class AutoBasketSpike extends LinearOpMode {
         public void Wait(double seconds) {
             Time.reset();
             while (Time.milliseconds()  < seconds * 1000) {
+
                 // doesn't need anything
+
             } // end of while loop
+
         } // end of public void Wait
+
     } // end of public class
