+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Autonomous/AutoFirstAscent.java
@@ -1,34 +1,4 @@
-
-/* Copyright (c) 2022 FIRST. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided that
- * the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this list
- * of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice, this
- * list of conditions and the following disclaimer in the documentation and/or
- * other materials provided with the distribution.
- *
- * Neither the name of FIRST nor the names of its contributors may be used to endorse or
- * promote products derived from this software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
- * LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.Autonomous;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
@@ -40,65 +10,23 @@ import com.qualcomm.robotcore.util.Range;
 
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
+import org.firstinspires.ftc.teamcode.Assemblies.LimbArm;
+import org.firstinspires.ftc.teamcode.Assemblies.SampleClaw;
 
-/*
- *  This OpMode illustrates the concept of driving an autonomous path based on Gyro (IMU) heading and encoder counts.
- *  The code is structured as a LinearOpMode
- *
- *  The path to be followed by the robot is built from a series of drive, turn or pause steps.
- *  Each step on the path is defined by a single function call, and these can be strung together in any order.
- *
- *  The code REQUIRES that you have encoders on the drive motors, otherwise you should use: RobotAutoDriveByTime;
- *
- *  This code uses the Universal IMU interface so it will work with either the BNO055, or BHI260 IMU.
- *  To run as written, the Control/Expansion hub should be mounted horizontally on a flat part of the robot chassis.
- *  The REV Logo should be facing UP, and the USB port should be facing forward.
- *  If this is not the configuration of your REV Control Hub, then the code should be modified to reflect the correct orientation.
- *
- *  This sample requires that the drive Motors have been configured with names : left_drive and right_drive.
- *  It also requires that a positive power command moves both motors forward, and causes the encoders to count UP.
- *  So please verify that both of your motors move the robot forward on the first move.  If not, make the required correction.
- *  See the beginning of runOpMode() to set the FORWARD/REVERSE option for each motor.
- *
- *  This code uses RUN_TO_POSITION mode for driving straight, and RUN_USING_ENCODER mode for turning and holding.
- *  Note: This code implements the requirement of calling setTargetPosition() at least once before switching to RUN_TO_POSITION mode.
- *
- *  Notes:
- *
- *  All angles are referenced to the coordinate-frame that is set whenever resetHeading() is called.
- *  In this sample, the heading is reset when the Start button is touched on the Driver Station.
- *  Note: It would be possible to reset the heading after each move, but this would accumulate steering errors.
- *
- *  The angle of movement/rotation is assumed to be a standardized rotation around the robot Z axis,
- *  which means that a Positive rotation is Counter Clockwise, looking down on the field.
- *  This is consistent with the FTC field coordinate conventions set out in the document:
- *  https://ftc-docs.firstinspires.org/field-coordinate-system
- *
- *  Control Approach.
- *
- *  To reach, or maintain a required heading, this code implements a basic Proportional Controller where:
- *
- *      Steering power = Heading Error * Proportional Gain.
- *
- *      "Heading Error" is calculated by taking the difference between the desired heading and the actual heading,
- *      and then "normalizing" it by converting it to a value in the +/- 180 degree range.
- *
- *      "Proportional Gain" is a constant that YOU choose to set the "strength" of the steering response.
- *
- *  Use Android Studio to Copy this Class, and Paste it into your "TeamCode" folder with a new name.
- *  Remove or comment out the @Disabled line to add this OpMode to the Driver Station OpMode list
- */
-
-@Autonomous(name="Robot: Lift test", group="Robot")
+@Autonomous(name="AutoFirstAscent", group="Robot")
 //@Disabled
-public class liftNearNet extends LinearOpMode {
+public class AutoFirstAscent extends LinearOpMode {
 
     /* Declare OpMode members. */
     private DcMotor leftFrontDrive   = null;
     private DcMotor rightFrontDrive = null;
     private DcMotor rightBackDrive = null;
     private DcMotor leftBackDrive = null;
-    private IMU             imu         = null;      // Control/Expansion Hub IMU
+    private IMU imu         = null;
+    private final ElapsedTime Time = new ElapsedTime();
+    LimbArm arm;
+    SampleClaw claw;
+    // Control/Expansion Hub IMU
 
     private double          headingError  = 0;
 
@@ -113,8 +41,6 @@ public class liftNearNet extends LinearOpMode {
     private int leftBackTarget = 0;
     private int rightBackTarget = 0;
     private int rightFrontTarget = 0;
-    private int     SecondleftTarget    = 0; //Only used for strafing
-    private int     SecondrightTarget   = 0; //
     // Calculate the COUNTS_PER_INCH for your specific drive train.
     // Go to your motor vendor website to determine your motor's COUNTS_PER_MOTOR_REV
     // For external drive gearing, set DRIVE_GEAR_REDUCTION as needed.
@@ -130,7 +56,7 @@ public class liftNearNet extends LinearOpMode {
     // These constants define the desired driving/control characteristics
     // They can/should be tweaked to suit the specific robot drive train.
     static final double     DRIVE_SPEED             = 0.7;     // Max driving speed for better distance accuracy.
-    static final double     TURN_SPEED              = 0.4;     // Max turn speed to limit turn rate.
+    static final double     TURN_SPEED              = 0.5;     // Max turn speed to limit turn rate.
     static final double     HEADING_THRESHOLD       = 2.0 ;    // How close must the heading get to the target before moving to next step.
     // Requiring more accuracy (a smaller number) will often make the turn take longer to get into the final position.
     // Define the Proportional control coefficient (or GAIN) for "heading control".
@@ -143,9 +69,9 @@ public class liftNearNet extends LinearOpMode {
     public void Orientations() {
         leftFrontDrive.setDirection(DcMotor.Direction.REVERSE);
         leftBackDrive.setDirection(DcMotor.Direction.REVERSE);
-        rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-        rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
-
+        rightFrontDrive.setDirection(DcMotor.Direction.FORWARD);
+        rightBackDrive.setDirection(DcMotor.Direction.FORWARD);
+        Wait(0.1);
     }
     public void flip(String FirstLocation, String SecondLocation) {
         // The input is expected to be L or R for the first value and F or B for the second value
@@ -163,27 +89,33 @@ public class liftNearNet extends LinearOpMode {
                     leftBackDrive.setDirection(DcMotor.Direction.REVERSE);
                 }
             }
-        } else {
+        }
+        else {
             if (SecondLocation.equals("F")) {
                 if (rightFrontDrive.getDirection() == DcMotor.Direction.REVERSE) {
                     rightFrontDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
+                } // end of if statement
+                else {
                     rightFrontDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            } else {
+                } // end of else statement
+            } // end of if statement
+            else {
                 if (rightBackDrive.getDirection() == DcMotor.Direction.REVERSE) {
                     rightBackDrive.setDirection(DcMotor.Direction.FORWARD);
-                } else {
+                } // end of if statement
+                else {
                     rightBackDrive.setDirection(DcMotor.Direction.REVERSE);
-                }
-            }
-
-        }
+                } // end of else Statement
+            } // end of else Statement
+        } // end of else statement
+        Wait(0.1);
     }
-
 
     @Override
     public void runOpMode() {
+
+        arm = new LimbArm(hardwareMap, telemetry);
+        claw = new SampleClaw(hardwareMap);
 
         // Initialize the drive system variables.
         leftFrontDrive  = hardwareMap.get(DcMotor.class, "leftFront");
@@ -199,7 +131,7 @@ public class liftNearNet extends LinearOpMode {
          *
          * To Do:  EDIT these two lines to match YOUR mounting configuration.
          */
-        RevHubOrientationOnRobot.LogoFacingDirection logoDirection = RevHubOrientationOnRobot.LogoFacingDirection.LEFT;
+        RevHubOrientationOnRobot.LogoFacingDirection logoDirection = RevHubOrientationOnRobot.LogoFacingDirection.RIGHT;
         RevHubOrientationOnRobot.UsbFacingDirection  usbDirection  = RevHubOrientationOnRobot.UsbFacingDirection.UP;
         RevHubOrientationOnRobot orientationOnRobot = new RevHubOrientationOnRobot(logoDirection, usbDirection);
 
@@ -222,22 +154,37 @@ public class liftNearNet extends LinearOpMode {
         leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         imu.resetYaw();
+
         // Wait for the game to start (Display Gyro value while waiting)
+
         while (opModeInInit()) {
             telemetry.addData(">", "Robot Heading = %4.0f", getHeading());
             telemetry.update();
-        }
+        } // end of while loop
 
         // Set the encoders for closed loop speed control, and reset the heading.
-        // BEGIN AUTO CODE //
-        driveStraight(DRIVE_SPEED, 4, getHeading());
-        turnToHeading(TURN_SPEED, 90);
-//        //extend arm and then drop sample and pull it back down.
-        turnToHeading(TURN_SPEED,0 );
-        driveStraight(DRIVE_SPEED, 50, getHeading());
+        // BEGIN AUTO CODE
+
+        StrafeRobot(TURN_SPEED, -6, 0);
+        // Wait(1);
+        driveStraight(TURN_SPEED, 62, 0);
+        // Wait(1);
         turnToHeading(TURN_SPEED, -90);
-        driveStraight(DRIVE_SPEED, 15, getHeading());
-        //extend arm and drop onto bar
+        // Wait(1);
+        driveStraight(TURN_SPEED, 10, -90);
+         Wait(1);
+       // arm.armRotateAuto(-50);
+       // Wait(1);
+       // arm.AutoExtendMotor(500);
+        // Wait(1);
+       // arm.armRotateAuto(100);
+       // Wait(1);
+
+
+
+
+        // 0 for heading is whatever the robot's original position is. 90 goes to the left of 0, and negative 90 goes to the right of the 0.
+        // Lastly, 180 goes the exact opposite of where 0 is.
         // Step through each leg of the path,
         // Notes:   Reverse movement is obtained by setting a negative distance (not speed)
         //          holdHeading() is used after turns to let the heading stabilize
@@ -245,7 +192,8 @@ public class liftNearNet extends LinearOpMode {
         telemetry.addData("Path", "Complete");
         telemetry.update();
         sleep(1000);  // Pause to display last telemetry message.
-    }
+
+    } // end of public void runOpMode
 
     /*
      * ====================================================================================================
@@ -267,9 +215,8 @@ public class liftNearNet extends LinearOpMode {
      *                   0 = fwd. +ve is CCW from fwd. -ve is CW from forward.
      *                   If a relative angle is required, add/subtract from the current robotHeading.
      */
-    public void driveStraight(double maxDriveSpeed,
-                              double distance,
-                              double heading) {
+    public void driveStraight(double maxDriveSpeed, double distance, double heading) {
+
         Orientations();
 
         // Ensure that the OpMode is still active
@@ -285,13 +232,10 @@ public class liftNearNet extends LinearOpMode {
             // Set Target FIRST, then turn on RUN_TO_POSITION
             // If Strafing then reverse motor directions
 
-
             leftFrontDrive.setTargetPosition(leftFrontTarget);
             rightFrontDrive.setTargetPosition(rightFrontTarget);
             leftBackDrive.setTargetPosition(leftBackTarget);
             rightBackDrive.setTargetPosition(rightBackTarget);
-
-
 
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
@@ -319,7 +263,7 @@ public class liftNearNet extends LinearOpMode {
 
                 // Display drive status for the driver.
                 sendTelemetry(true);
-            }
+            } // end of while loop
 
             // Stop all motion & Turn off RUN_TO_POSITION
             moveRobot(0, 0);
@@ -327,18 +271,27 @@ public class liftNearNet extends LinearOpMode {
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        }
-    }
+        } // end of if statement
+    } // end of public void driveStraight
     public void StrafeRobot(double maxDriveSpeed, double distance, double heading) {
         Orientations();
-        flip("L", "F");
-        flip("R", "B");
-//    telemetry.addData("Encoder Count aim", (distance * COUNTS_PER_INCH));
-        if (opModeIsActive()) {
+        if (distance > 0) {
+            flip("L", "B");
+            flip("R", "F");
+        } else {
+            flip("L", "F");
+            flip("R", "B");
+        }
 
+        telemetry.addData("Encoder Count aim", (distance * COUNTS_PER_INCH));
+        if (opModeIsActive()) {
+            leftFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+            leftBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+            rightFrontDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+            rightBackDrive.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
             // Determine new target position, and pass to motor controller
             // Determine new target position, and pass to motor controller
-            int moveCounts = (int)(distance * COUNTS_PER_INCH);
+            int moveCounts = (int)(Math.abs(distance) * COUNTS_PER_INCH);
             leftFrontTarget = leftFrontDrive.getCurrentPosition() + moveCounts;
             leftBackTarget = leftBackDrive.getCurrentPosition() + moveCounts;
             rightFrontTarget = rightFrontDrive.getCurrentPosition() + moveCounts;
@@ -346,14 +299,10 @@ public class liftNearNet extends LinearOpMode {
 
             // Set Target FIRST, then turn on RUN_TO_POSITION
             // If Strafing then reverse motor directions
-
-
             leftFrontDrive.setTargetPosition(leftFrontTarget);
             rightFrontDrive.setTargetPosition(rightFrontTarget);
             leftBackDrive.setTargetPosition(leftBackTarget);
             rightBackDrive.setTargetPosition(rightBackTarget);
-
-
 
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_TO_POSITION);
@@ -373,32 +322,32 @@ public class liftNearNet extends LinearOpMode {
                 turnSpeed = getSteeringCorrection(heading, P_DRIVE_GAIN);
 
                 // if driving in reverse, the motor correction also needs to be reversed
-                if (distance < 0)
-                    turnSpeed *= -1.0;
-
+//            if (distance < 0) {
+//                turnSpeed *= -1.0;
+//            }
                 // Apply the turning correction to the current driving speed.
-                moveRobot(driveSpeed, turnSpeed);
+                RotateRoboto(driveSpeed, turnSpeed, distance < 0);
 
                 // Display drive status for the driver.
                 // sendTelemetry(true);
             }
-//        telemetry.addData(" LF end encoder counter", leftFrontDrive.getCurrentPosition() );
-//        telemetry.addData(" LB end encoder counter", leftBackDrive.getCurrentPosition() );
-//        telemetry.addData(" RF end encoder counter", rightFrontDrive.getCurrentPosition() );
-//        telemetry.addData(" RB end encoder counter", rightBackDrive.getCurrentPosition() );
+            telemetry.addData(" LF end encoder counter", leftFrontDrive.getCurrentPosition() );
+            telemetry.addData(" LB end encoder counter", leftBackDrive.getCurrentPosition() );
+            telemetry.addData(" RF end encoder counter", rightFrontDrive.getCurrentPosition() );
+            telemetry.addData(" RB end encoder counter", rightBackDrive.getCurrentPosition() );
             telemetry.update();
-//        sleep(100000);
 
             // Stop all motion & Turn off RUN_TO_POSITION
-            // turnToHeading(maxDriveSpeed,heading);
             moveRobot(0, 0);
             leftFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             rightFrontDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             leftBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             rightBackDrive.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+            Orientations();
+            turnToHeading(maxDriveSpeed,heading);
         }
 
-    }
+    } // end of public void StrafeRobot
 
     /**
      *  Spin on the central axis to point in a new direction.
@@ -415,12 +364,13 @@ public class liftNearNet extends LinearOpMode {
      *              If a relative angle is required, add/subtract from current heading.
      */
     public void turnToHeading(double maxTurnSpeed, double heading) {
-
+        //  Orientations();
+        double TempThreshold = 1;
         // Run getSteeringCorrection() once to pre-calculate the current error
         getSteeringCorrection(heading, P_DRIVE_GAIN);
 
         // keep looping while we are still active, and not on heading.
-        while (opModeIsActive() && (Math.abs(headingError) > HEADING_THRESHOLD)) {
+        while (opModeIsActive() && (Math.abs(headingError) > TempThreshold)) {
 
             // Determine required steering to keep on heading
             turnSpeed = getSteeringCorrection(heading, P_TURN_GAIN);
@@ -431,14 +381,14 @@ public class liftNearNet extends LinearOpMode {
             // Pivot in place by applying the turning correction
             moveRobot(0, turnSpeed);
 
-            telemetry.addData("HeadingErr", headingError);
-            telemetry.addData("HeadinggetThresh",HEADING_THRESHOLD);
-            telemetry.update();
-        }
+//            telemetry.addData("HeadingErr", headingError);
+//            telemetry.addData("HeadinggetThresh",TempThreshold);
+//            telemetry.update();
+        } // end of while loop
 
         // Stop all motion;
         moveRobot(0, 0);
-    }
+    } // end of public void turnToHeading
 
     /**
      *  Obtain & hold a heading for a finite amount of time
@@ -471,11 +421,11 @@ public class liftNearNet extends LinearOpMode {
 
             // Display drive status for the driver.
             sendTelemetry(false);
-        }
+        } // end of while loop
 
         // Stop all motion;
         moveRobot(0, 0);
-    }
+    } // end of public void holdHeading
 
     // **********  LOW Level driving functions.  ********************
 
@@ -498,7 +448,7 @@ public class liftNearNet extends LinearOpMode {
 
         // Multiply the error by the gain to determine the required steering correction/  Limit the result to +/- 1.0
         return Range.clip(headingError * proportionalGain, -1, 1);
-    }
+    } // end of public double getSteeringCorrection
 
     /**
      * Take separate drive (fwd/rev) and turn (right/left) requests,
@@ -515,16 +465,41 @@ public class liftNearNet extends LinearOpMode {
 
         // Scale speeds down if either one exceeds +/- 1.0;
         double max = Math.max(Math.abs(leftSpeed), Math.abs(rightSpeed));
-        if (max > 1.0)
-        {
+        if (max > 1.0) {
             leftSpeed /= max;
             rightSpeed /= max;
-        }
+        } // end of if statement
 
-        leftFrontDrive.setPower(leftSpeed);
+        leftFrontDrive.setPower(leftSpeed * 0.85);
         rightFrontDrive.setPower(rightSpeed);
         leftBackDrive.setPower(leftSpeed);
         rightBackDrive.setPower(rightSpeed);
+    } // end of public void moveRobot
+
+    public void RotateRoboto(double drive, double turn, Boolean isLeft ) {
+//        driveSpeed = drive;     // save this value as a class member so it can be used by telemetry.
+//        turnSpeed  = turn;      // save this value as a class member so it can be used by telemetry.
+//        leftSpeed  = drive - turn;
+//        rightSpeed = drive + turn;
+        leftFrontDrive.setPower(drive + ((isLeft ? 1 : -1) * turn));
+        rightFrontDrive.setPower(drive + ((isLeft ? 1 : -1) * turn));
+        leftBackDrive.setPower(drive + ((isLeft ? -1 : 1) * turn));
+        rightBackDrive.setPower(drive + ((isLeft ? -1 : 1) * turn));
+
+        // Scale speeds down if either one exceeds +/- 1.0;
+//        double max = Math.max(Math.abs(leftSpeed), Math.abs(rightSpeed));
+//        if (max > 1.0)
+//        {
+//            leftSpeed /= max;
+//            rightSpeed /= max;
+//        }
+
+
+
+//        leftFrontDrive.setPower((isLeft ? -1 : 1) * leftSpeed);
+//        rightFrontDrive.setPower((isLeft ? 1 : -1) * rightSpeed);
+//        leftBackDrive.setPower((isLeft ? 1 : -1) * leftSpeed);
+//        rightBackDrive.setPower((isLeft ? -1 : 1) * rightSpeed);
     }
 
     /**
@@ -540,20 +515,31 @@ public class liftNearNet extends LinearOpMode {
             telemetry.addData("Target Pos L:R",  "%7d:%7d",      leftFrontTarget, rightFrontTarget);
             telemetry.addData("Actual Pos L:R",  "%7d:%7d:%7d:%7d",      leftFrontDrive.getCurrentPosition(),
                     rightFrontDrive.getCurrentPosition(),leftBackDrive.getCurrentPosition(),rightBackDrive.getCurrentPosition());
-        } else {
+        } //end of if statement
+
+        else {
             telemetry.addData("Motion", "Turning");
-        }
+        } // end of else statement
 
         telemetry.addData("Heading- Target : Current", "%5.2f : %5.0f", targetHeading, getHeading());
         telemetry.addData("Error  : Steer Pwr",  "%5.1f : %5.1f", headingError, turnSpeed);
         telemetry.addData("Wheel Speeds L : R", "%5.2f : %5.2f", leftSpeed, rightSpeed);
         telemetry.update();
-    }
+
+    } // end of private void sendTelemetry
     /**
      * read the Robot heading directly from the IMU (in degrees)
      */
     public double getHeading() {
         YawPitchRollAngles orientation = imu.getRobotYawPitchRollAngles();
         return orientation.getYaw(AngleUnit.DEGREES);
-    }
-}
+    } // end of public double getHeading
+
+    public void Wait(double seconds) {
+        Time.reset();
+        while (Time.milliseconds()  < seconds * 1000) {
+            // doesn't need anything
+        } // end of while loop
+    } // end of public void Wait
+} // end of public class
+
