--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Robot.java
@@ -1,21 +1,0 @@
-// This file is a system file.
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.hardware.HardwareMap;
-
-import org.firstinspires.ftc.robotcontroller.external.samples.UtilityOctoQuadConfigMenu;
-import org.firstinspires.ftc.robotcore.external.Telemetry;
-
-public class Robot {
-    public DriveTrain driveTrain;
-    //public AscentMechanism ascentMechanism;
-    public SampleClaw sampleClaw;
-    public LimbArm limbArm;
-    // This combines all the subsystems.
-    public Robot(HardwareMap hwMap, Telemetry telemetry) {
-        driveTrain = new DriveTrain(hwMap, telemetry);
-   //     ascentMechanism = new AscentMechanism(hwMap);
-        sampleClaw = new SampleClaw(hwMap);
-        limbArm = new LimbArm(hwMap, telemetry);
-    }
-}
